# Optimize MQA computation.