--- a/pyproject.toml
+++ b/pyproject.toml
@@ -22,6 +22,7 @@ dependencies = [
 core = [
     "absl-py==2.1.0",
     "chex==0.1.86",  # chex 0.1.86 is required for jax 0.4.25.
+    "einops==0.8.0",
     "importlab==0.7",  # breaks pytype on 0.8
     "jax==0.4.34",
     "jaxlib==0.4.34",
@@ -53,10 +54,10 @@ apple-silicon = [
 ]
 # Requirements for testing and development.
 dev = [
+    "axlearn[core]",  # core
     "axlearn[audio]",  # audio tests
     "axlearn[orbax]",  # checkpointer tests
     "black==23.1a1",  # formatting
-    "einops==0.8.0",
     "evaluate",
     "isort",  # formatting
     "pika==1.3.2",  # used by event queue